feat: add organisation settings tab
